// src/lib/api/models/createCommissionPlanDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */
import type { CommissionRuleDto } from './commissionRuleDto';

export interface CreateCommissionPlanDto {
  name: string;
  description?: string;
  version?: number;
  priority?: number;
  effectiveFrom: string;
  effectiveTo?: string;
  rules: CommissionRuleDto[];
}